Move `RequestInfo` class to `network`
