+++ search-api/src/main/scala/no/ndla/searchapi/controller/InternController.scala
@@ -7,6 +7,8 @@
 
 package no.ndla.searchapi.controller
 
+import no.ndla.network.model.RequestInfo
+
 import java.util.concurrent.{Executors, TimeUnit}
 import no.ndla.searchapi.SearchApiProperties
 import no.ndla.searchapi.integration.{GrepApiClient, TaxonomyApiClient}
@@ -14,7 +16,7 @@ import no.ndla.searchapi.model.api.InvalidIndexBodyException
 import no.ndla.searchapi.model.domain.article.Article
 import no.ndla.searchapi.model.domain.draft.Draft
 import no.ndla.searchapi.model.domain.learningpath._
-import no.ndla.searchapi.model.domain.{Content, ReindexResult, RequestInfo}
+import no.ndla.searchapi.model.domain.{Content, ReindexResult}
 import no.ndla.searchapi.service.search.{ArticleIndexService, DraftIndexService, IndexService, LearningPathIndexService}
 import org.scalatra._
