+++ search-api/src/main/scala/no/ndla/searchapi/service/search/MultiSearchService.scala
@@ -14,6 +14,7 @@ import com.sksamuel.elastic4s.requests.searches.queries.compound.BoolQuery
 import com.typesafe.scalalogging.LazyLogging
 import no.ndla.language.Language.AllLanguages
 import no.ndla.language.model.Iso639
+import no.ndla.network.model.RequestInfo
 import no.ndla.search.Elastic4sClient
 import no.ndla.searchapi.SearchApiProperties
 import no.ndla.searchapi.SearchApiProperties.{
@@ -23,7 +24,7 @@ import no.ndla.searchapi.SearchApiProperties.{
 }
 import no.ndla.searchapi.model.api.ResultWindowTooLargeException
 import no.ndla.searchapi.model.domain.article.Availability
-import no.ndla.searchapi.model.domain.{RequestInfo, SearchResult}
+import no.ndla.searchapi.model.domain.SearchResult
 import no.ndla.searchapi.model.search.SearchType
 import no.ndla.searchapi.model.search.settings.SearchSettings
